SDK 4.0 updates to BuildPlayer
No longer supporting < 2018

+++ Assets/Editor/BuildPlayer.cs
@@ -6,14 +6,10 @@ using UnityEngine;
 using UnityEditor;
 using UnityEditor.Callbacks;
 
-#if !TEAK_NOT_AVAILABLE
-using TeakEditor.iOS.Xcode;
-using TeakEditor.iOS.Xcode.Extensions;
-#endif
+using UnityEditor.iOS.Xcode;
+using UnityEditor.iOS.Xcode.Extensions;
 
-#if UNITY_2018_1_OR_NEWER
 using UnityEditor.Build.Reporting;
-#endif
 
 using GooglePlayServices;
 
@@ -93,34 +89,20 @@ public class BuildPlayer
     }
 
     static void DoBuildPlayer(BuildPlayerOptions buildPlayerOptions) {
-#if UNITY_2018_1_OR_NEWER
         BuildReport report = BuildPipeline.BuildPlayer(buildPlayerOptions);
         if(report.summary.result == BuildResult.Succeeded) {
             // Preserve the merged AndroidManifest.xml
             if (buildPlayerOptions.target == BuildTarget.Android) {
                 string mergedManifestPath = System.IO.Path.GetFullPath(Application.dataPath + "/../Temp/StagingArea/AndroidManifest.xml");
                 string outputManifestPath = System.IO.Path.GetFullPath(Application.dataPath + "/../AndroidManifest.merged.xml");
-                System.IO.File.Copy(mergedManifestPath, outputManifestPath, true);
+                if (System.IO.File.Exists(mergedManifestPath)) {
+                    System.IO.File.Copy(mergedManifestPath, outputManifestPath, true);
+                }
             }
             EditorApplication.Exit(0);
         } else if(report.summary.result == BuildResult.Failed) {
             EditorApplication.Exit(1);
         }
-#else
-        string error = BuildPipeline.BuildPlayer(buildPlayerOptions);
-        if (string.IsNullOrEmpty(error)) {
-            // Preserve the merged AndroidManifest.xml
-            if (buildPlayerOptions.target == BuildTarget.Android) {
-                string mergedManifestPath = System.IO.Path.GetFullPath(Application.dataPath + "/../Temp/StagingArea/AndroidManifest.xml");
-                string outputManifestPath = System.IO.Path.GetFullPath(Application.dataPath + "/../AndroidManifest.merged.xml");
-                System.IO.File.Copy(mergedManifestPath, outputManifestPath, true);
-            }
-
-            EditorApplication.Exit(0);
-        } else {
-            EditorApplication.Exit(1);
-        }
-#endif
     }
 
     static void WebGL()
